refactor: lib_plot bugfix: fix resolution fit range
1. pack plot functions from util_lib into lib_plot;
2. resolution fit only in range of acceptance.
