+++ lib_plot/.python-version
@@ -1,0 +1,1 @@
+3.8.16
