+++ lib_plot/README.md
@@ -1,0 +1,16 @@
+# lib-plot
+---
+## INFO:
+
+Module: GRID Plot Lib
+Author: Wqd
+
+GRID Collaboration/Calibration
+
+---
+
+## Features:
+
++ plot count-adc raw & fit graph
++ plot temp-bias fit graph
++ plot energy-channel fit graph
