+++ lib_plot/pyproject.toml
@@ -1,0 +1,21 @@
+[project]
+name = "lib-plot"
+version = "0.1.0"
+description = "Add your description here"
+authors = [
+    { name = "wqd", email = "2798730887@qq.com" }
+]
+dependencies = []
+readme = "README.md"
+requires-python = ">= 3.8"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.rye]
+managed = true
+dev-dependencies = []
+
+[tool.hatch.metadata]
+allow-direct-references = true
