+++ lib_plot/src/lib_plot/__init__.py
@@ -1,0 +1,1 @@
+from .plot import raw_plot, fit_plot_single_channel, fit_err_plot_2d, ec_plot
