+++ pyproject.toml
@@ -7,6 +7,7 @@ authors = [
 ]
 dependencies = [
     "lib_reader",
+    "lib_plot",
     "numpy>=1.24.4",
     "matplotlib>=3.7.2",
     "scipy>=1.10.1",
